+++ code/src/sixsq/nuvla/auth/password.clj
@@ -2,11 +2,9 @@
   (:refer-clojure :exclude [update])
   (:require
     [buddy.hashers :as hashers]
-    [clojure.string :as str]
     [sixsq.nuvla.db.filter.parser :as parser]
     [sixsq.nuvla.db.impl :as db]
     [sixsq.nuvla.server.resources.common.crud :as crud]
-    [sixsq.nuvla.server.resources.group :as group]
     [sixsq.nuvla.server.resources.user-identifier :as user-identifier]))
 
 
@@ -76,23 +74,3 @@
                               :hash)]
     (when (valid-password? password password-hash)
       user)))
-
-
-(defn collect-groups-for-user
-  [id]
-  (let [group-set (->> (crud/query-as-admin
-                         group/resource-type
-                         {:cimi-params {:filter (parser/parse-cimi-filter (format "users='%s'" id))
-                                        :select ["id"]}})
-                       second
-                       (map :id)
-                       (cons "group/nuvla-user")            ;; if there's an id, then the user is authenticated
-                       (cons "group/nuvla-anon")            ;; all users are in the nuvla-anon pseudo-group
-                       set)]
-    (str/join " " (sort group-set))))
-
-
-(defn create-claims
-  [{:keys [id] :as user}]
-  {:user-id id
-   :claims  (collect-groups-for-user id)})