+++ code/test/sixsq/nuvla/server/resources/session_password_lifecycle_test.clj
@@ -5,11 +5,16 @@
     [clojure.test :refer [deftest is use-fixtures]]
     [peridot.core :refer [content-type header request session]]
     [postal.core :as postal]
+    [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.auth.utils.sign :as sign]
     [sixsq.nuvla.server.app.params :as p]
-    [sixsq.nuvla.server.middleware.authn-info :refer [authn-cookie authn-info-header]]
+    [sixsq.nuvla.server.middleware.authn-info
+     :refer [authn-cookie authn-info-header wrap-authn-info]]
     [sixsq.nuvla.server.resources.email.utils :as email-utils]
+    [sixsq.nuvla.server.resources.group :as group]
+    [sixsq.nuvla.server.resources.group-template :as group-tpl]
     [sixsq.nuvla.server.resources.lifecycle-test-utils :as ltu]
+    [sixsq.nuvla.server.resources.nuvlabox :as nuvlabox]
     [sixsq.nuvla.server.resources.session :as session]
     [sixsq.nuvla.server.resources.session-template :as st]
     [sixsq.nuvla.server.resources.user :as user]
@@ -157,7 +162,8 @@
         (is (= jane-user-id (:user-id authn-info)))
         (is (= #{"group/nuvla-user"
                  "group/nuvla-anon"
-                 uri}
+                 uri
+                 jane-user-id}
                (-> authn-info
                    :claims
                    (str/split #"\s")
@@ -200,7 +206,8 @@
             (ltu/is-status 200)
             (ltu/is-id id)
             (ltu/is-operation-present :delete)
-            (ltu/is-operation-absent :edit))
+            (ltu/is-operation-absent :edit)
+            (ltu/is-operation-absent :claim))
 
         ; check contents of session
         (let [{:keys [name description tags]} (-> session-user
@@ -273,4 +280,182 @@
           (ltu/is-status 403)))
 
     ))
+
+(deftest claim-account-test
+
+  (let [app                (ltu/ring-app)
+        session-json       (content-type (session app) "application/json")
+        session-anon       (header session-json authn-info-header "user/unknown group/nuvla-anon")
+        session-admin      (header session-json authn-info-header "user/super group/nuvla-admin group/nuvla-user group/nuvla-anon")
+
+        href               (str st/resource-type "/password")
+
+        username           "user/bob"
+        plaintext-password "BobBob-0"
+        user-id            (create-user session-admin
+                                        :username username
+                                        :password plaintext-password
+                                        :activated? true
+                                        :email "bob@example.org")]
+
+
+    ;; anon with valid activated user can create session
+    (let [username           "user/bob"
+          plaintext-password "BobBob-0"
+
+          valid-create       {:template {:href     href
+                                         :username username
+                                         :password plaintext-password}}]
+
+      ; anonymous create must succeed
+      (let [session-1          (-> session-anon
+                                   (request base-uri
+                                            :request-method :post
+                                            :body (json/write-str valid-create))
+                                   (ltu/body->edn)
+                                   (ltu/is-set-cookie)
+                                   (ltu/is-status 201))
+            session-1-id       (ltu/body-resource-id session-1)
+
+            session-1-uri      (ltu/location session-1)
+            sesssion-1-abs-uri (str p/service-context session-1-uri)
+            handler            (wrap-authn-info identity)
+            authn-session-1    (-> {:cookies (get-in session-1 [:response :cookies])}
+                                   handler
+                                   seq
+                                   flatten)
+
+            group-identifier   "alpha"
+            group-alpha        (str group/resource-type "/" group-identifier)
+            group-create       {:template {:href             (str group-tpl/resource-type "/generic")
+                                           :group-identifier group-identifier}}
+            group-url          (-> session-admin
+                                   (request (str p/service-context group/resource-type)
+                                            :request-method :post
+                                            :body (json/write-str group-create))
+                                   (ltu/body->edn)
+                                   (ltu/is-status 201)
+                                   (ltu/location-url))]
+
+        ; user without additional group should not have operation claim
+        (-> (apply request session-json (concat [sesssion-1-abs-uri] authn-session-1))
+            (ltu/body->edn)
+            (ltu/is-status 200)
+            (ltu/is-id session-1-id)
+            (ltu/is-operation-present :delete)
+            (ltu/is-operation-absent :edit)
+            (ltu/is-operation-absent :claim))
+
+        ;; add user to group/alpha
+        (-> session-admin
+            (request group-url
+                     :request-method :put
+                     :body (json/write-str {:users [user-id]}))
+            (ltu/body->edn)
+            (ltu/is-status 200))
+
+        ;; check claim operation present
+        (let [session-2         (-> session-anon
+                                    (request base-uri
+                                             :request-method :post
+                                             :body (json/write-str valid-create))
+                                    (ltu/body->edn)
+                                    (ltu/is-set-cookie)
+                                    (ltu/is-status 201))
+              session-2-id      (ltu/body-resource-id session-2)
+
+              session-2-uri     (ltu/location session-2)
+              session-2-abs-uri (str p/service-context session-2-uri)
+
+              authn-session-2   (-> {:cookies (get-in session-2 [:response :cookies])}
+                                    handler
+                                    seq
+                                    flatten)
+
+              claim-op-url      (-> (apply request session-json (concat [session-2-abs-uri] authn-session-2))
+                                    (ltu/body->edn)
+                                    (ltu/is-status 200)
+                                    (ltu/get-op-url :claim))]
+
+          (-> (apply request session-json (concat [session-2-abs-uri] authn-session-2))
+              (ltu/body->edn)
+              (ltu/is-status 200)
+              (ltu/is-id session-2-id)
+              (ltu/is-operation-present :delete)
+              (ltu/is-operation-absent :edit)
+              (ltu/is-operation-present :claim))
+
+          ;; claiming group-beta should fail
+          (-> (apply request session-json
+                     (concat [claim-op-url :body (json/write-str {:claim "group/beta"})
+                              :request-method :post] authn-session-2))
+              (ltu/body->edn)
+              (ltu/is-status 403)
+              (ltu/message-matches #"Switch account cannot be done to requested claim:.*"))
+
+          ;; claim with old session fail because wasn't in group/alpha
+          (-> (apply request session-json
+                     (concat [claim-op-url :body (json/write-str {:claim group-alpha})
+                              :request-method :post] authn-session-1))
+              (ltu/body->edn)
+              (ltu/is-status 403))
+
+          ;; claim group-alpha
+          (let [cookie-claim        (-> (apply request session-json
+                                               (concat [claim-op-url :body (json/write-str
+                                                                             {:claim group-alpha})
+                                                        :request-method :post] authn-session-2))
+                                        (ltu/body->edn)
+                                        (ltu/is-status 200)
+                                        (ltu/is-set-cookie)
+                                        :response
+                                        :cookies)
+                authn-session-claim (-> {:cookies cookie-claim}
+                                        handler
+                                        seq
+                                        flatten)]
+            (is (= group-alpha
+                   (-> {:cookies cookie-claim}
+                       handler
+                       auth/current-authentication
+                       :user-id)))
+
+            (-> (apply request session-json (concat [session-2-abs-uri] authn-session-claim))
+                (ltu/body->edn)
+                (ltu/is-status 200)
+                (ltu/is-id session-2-id)
+                (ltu/is-operation-present :delete)
+                (ltu/is-operation-absent :edit)
+                (ltu/is-operation-present :claim))
+
+            ;; try create NuvlaBox and check who is the owner
+            (let [nuvlabox-url (-> (apply request session-json
+                                          (concat [(str p/service-context nuvlabox/resource-type)
+                                                   :body (json/write-str {})
+                                                   :request-method :post] authn-session-claim))
+                                   (ltu/body->edn)
+                                   (ltu/is-status 201)
+                                   (ltu/location-url))]
+
+              (-> (apply request session-json (concat [nuvlabox-url] authn-session-claim))
+                  (ltu/body->edn)
+                  (ltu/is-status 200)
+                  (ltu/is-key-value :owner group-alpha)))
+
+            (let [cookie-claim-back (-> (apply request session-json
+                                               (concat [claim-op-url :body (json/write-str
+                                                                             {:claim user-id})
+                                                        :request-method :post] authn-session-claim))
+                                        (ltu/body->edn)
+                                        (ltu/is-status 200)
+                                        (ltu/is-set-cookie)
+                                        :response
+                                        :cookies)]
+              (is (= user-id
+                     (-> {:cookies cookie-claim-back}
+                         handler
+                         auth/current-authentication
+                         :user-id))))))
+
+        ))))